requirements.txt
requests
beautifulsoup4 
fake_useragent
reportlab
streamlit
google-generativeai
python-dotenv
langchain
chromadb
faiss-cpu
langchain_google_genai
langchain-community